Warning!!! Requires rebuild. Fixed bug in LightConnes::ASCII_XMR::fastplanes_parallel() and in LightConnes::ASCII_XMRRT::fastplanes_parallel() Moved some things around for organization.
